requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.10.5
et_xmlfile==2.0.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
magic-filter==1.0.12
multidict==6.7.0
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
propcache==0.4.1
pydantic==2.11.10
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
yarl==1.22.0
python-docx==1.2.0
